# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "API Starwars",
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://localhost/api"
        },
        {
            "url": "http://127.0.0.1:8000/api"
        }
    ],
    "paths": {
...
        "/users": {
            "get": {
                "tags": [
                    "Users"
                ],
                "summary": "Query all users",
                "operationId": "index",
                "responses": {
                    "200": {
                        "description": "returns the registered users"
                    },
                    "422": {
                        "description": "validation error"
                    },
                    "500": {
                        "description": "internal error"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "Users"
                ],
                "summary": "register users",
                "operationId": "store",
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "name": {
                                        "type": "string"
                                    },
                                    "email": {
                                        "type": "string"
                                    },
                                    "password": {
                                        "oneOf": [
                                            {
                                                "type": "string"
                                            },
                                            {
                                                "type": "integer"
                                            }
                                        ]
                                    }
                                },
                                "type": "object",
                                "example": {
                                    "name": "Angel",
                                    "email": "angelmserranog@gmail.com",
                                    "password": 12345678
                                }
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "returns the registered object"
                    },
                    "422": {
                        "description": "validation error"
                    },
                    "500": {
                        "description": "internal error"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/users/{id}": {
            "get": {
                "tags": [
                    "Users"
                ],
                "summary": "Return a query",
                "operationId": "show",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "returns a registered query"
                    },
                    "422": {
                        "description": "validation error"
                    },
                    "500": {
                        "description": "internal error"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "put": {
                "tags": [
                    "Users"
                ],
                "summary": "Update a data",
                "operationId": "update",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "name": {
                                        "type": "string"
                                    },
                                    "email": {
                                        "type": "string"
                                    },
                                    "password": {
                                        "oneOf": [
                                            {
                                                "type": "string"
                                            },
                                            {
                                                "type": "integer"
                                            }
                                        ]
                                    }
                                },
                                "type": "object",
                                "example": {
                                    "name": "Angel",
                                    "email": "angelmserranog@gmail.com",
                                    "password": 12345678
                                }
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "returns a updated query"
                    },
                    "422": {
                        "description": "validation error"
                    },
                    "500": {
                        "description": "internal error"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "delete": {
                "tags": [
                    "Users"
                ],
                "summary": "Delete to data",
                "operationId": "destroy",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "returns a delete query"
                    },
                    "422": {
                        "description": "validation error"
                    },
                    "500": {
                        "description": "internal error"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "name": "bearerAuth",
                "in": "header",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    }
}